+++ src/app-components/TimePicker/utils/segmentTyping.ts
@@ -1,5 +1,5 @@
-import type { SegmentType } from 'src/app-components/TimePicker/keyboardNavigation';
-import type { TimeFormat } from 'src/app-components/TimePicker/TimePicker';
+import type { TimeFormat } from 'src/app-components/TimePicker/components/TimePicker';
+import type { SegmentType } from 'src/app-components/TimePicker/utils/keyboardNavigation';
 
 export interface SegmentTypingResult {
   value: string;
